Create KPoP_SELS.py
For a given phylogenetic tree and protein structure simulate evolution. Algorithm: (1) Mutations are randomly introduced throughout a protein. (2) The change in stability of a protein is calculated using contact energies from a lookup table. (3) A probability of fixation is calculated and accepted in proportion to the change in stability. (4) The new sequence is recorded. (5) Move on to the next sequence once the expected number of mutations are acquired along a particular branch length.